MLFlow

Description

Haskell Bindings for MLFlow: https://www.mlflow.org/

Documentation

type TrackingURI

Arguments

= String
Base URL and Port to mlflow server

trackingURI :: String -> String -> TrackingURI

Hostname and port to TrackingURI

trackingURI' :: String -> Int -> TrackingURI

Hostname and port (as Int) to TrackingURI

type ExperimentID

Arguments

= String
Experiment ID

createExperiment :: TrackingURI -> String -> IO ExperimentID

Create an experiment with a name. Returns the ID of the newly created experiment. Validates that another experiment with the same name does not already exist and fails if another experiment with the same name already exists.

listExperiments :: TrackingURI -> IO [Experiment]

Get a list of all experiments.

getExperiment :: TrackingURI -> ExperimentID -> IO Experiment

Get metadata for an experiment. This method works on deleted experiments.

getExperimentByName :: TrackingURI -> String -> IO Experiment

Get metadata for an experiment. (By Name)

deleteExperiment :: TrackingURI -> ExperimentID -> IO (Response ByteString)

Mark an experiment and associated metadata, runs, metrics, params, and tags for deletion. If the experiment uses FileStore, artifacts associated with experiment are also deleted.

resoterExperiment :: TrackingURI -> ExperimentID -> IO (Response ByteString)

Restore an experiment marked for deletion. This also restores associated metadata, runs, metrics, params, and tags. If experiment uses FileStore, underlying artifacts associated with experiment are also restored.

updateExperiment :: TrackingURI -> ExperimentID -> String -> IO (Response ByteString)

Update experiment metadata.

type RunID

Arguments

= String
Run ID

createRun :: TrackingURI -> ExperimentID -> [Tag] -> IO Run

Create a new run within an experiment. A run is usually a single execution of a machine learning or data ETL pipeline. MLflow uses runs to track Param, Metric, and RunTag associated with a single execution.

deleteRun :: TrackingURI -> RunID -> IO (Response ByteString)

Mark a run for deletion.

restoreRun :: TrackingURI -> RunID -> IO (Response ByteString)

Restore a deleted run.

getRun :: TrackingURI -> RunID -> IO Run

Get metadata, metrics, params, and tags for a run. In the case where multiple metrics with the same key are logged for a run, return only the value with the latest timestamp. If there are multiple values with the latest timestamp, return the maximum of these values.

updateRun :: TrackingURI -> RunID -> RunStatus -> Maybe Int -> IO RunInfo

Update run metadata.

logMetric :: TrackingURI -> RunID -> String -> Float -> Int -> IO (Response ByteString)

Log a metric for a run. A metric is a key-value pair (string key, float value) with an associated timestamp. Examples include the various metrics that represent ML model accuracy. A metric can be logged multiple times.

logBatch :: TrackingURI -> RunID -> [Metric] -> [Param] -> [Tag] -> IO (Response ByteString)

Log a batch of metrics, params, and tags for a run. If any data failed to be persisted, the server will respond with an error (non-200 status code). In case of error (due to internal server error or an invalid request), partial data may be written.

logBatch' :: TrackingURI -> RunID -> Int -> Map String Float -> Map String String -> IO (Response ByteString)

Just for Convenience

logParam :: TrackingURI -> RunID -> String -> String -> IO (Response ByteString)

Log a param used for a run. A param is a key-value pair (string key, string value). Examples include hyperparameters used for ML model training and constant dates and values used in an ETL pipeline. A param can be logged only once for a run.

getMetricHistory :: TrackingURI -> RunID -> String -> IO [Metric]

Get a list of all values for the specified metric for a given run.

setExperimentTag :: TrackingURI -> ExperimentID -> String -> String -> IO (Response ByteString)

Set a tag on an experiment. Experiment tags are metadata that can be updated.

setRunTag :: TrackingURI -> RunID -> String -> String -> IO (Response ByteString)

Set a tag on a run. Tags are run metadata that can be updated during a run and after a run completes.

deleteTag :: TrackingURI -> RunID -> String -> IO (Response ByteString)

Delete a tag on a run. Tags are run metadata that can be updated during a run and after a run completes.