MLFlow.DataStructures

Description

MLFlow Bindings

Documentation

data Tag

Tags

Constructors

Tag

Fields

tagKey :: String

The tag key.

tagValue :: String

The tag value.

data LifeCycleStage

Current Life Cycle Stage of the Experiment

Constructors

Active

Deleted

type ExperimentTag = Tag

Tag for an experiment.

data Experiment

Experiment

Constructors

Experiment

Fields

experimentId :: String

Unique identifier for the experiment.

experimentName :: String

Human readable name that identifies the experiment.

artifactLocation :: FilePath

Location where artifacts for the experiment are stored.

lifecycleStage :: LifeCycleStage

Current life cycle stage of the experiment: “active” or “deleted”. Deleted experiments are not returned by APIs.

lastUpdateTime :: Maybe Int

Last update time

creationTime :: Maybe Int

Creation time

experimentTags :: Maybe [ExperimentTag]

Tags: Additional metadata key-value pairs.

data FileInfo

Metadata of a single artifact file or directory.

Constructors

FileInfo

Fields

path :: FilePath

Path relative to the root artifact directory run.

isDir :: Bool

Whether the path is a directory.

fileSize :: Int

Size in bytes. Unset for directories.

data Metric

Metric associated with a run, represented as a key-value pair.

Constructors

Metric

Fields

metricKey :: String

Key identifying this metric.

metricValue :: Float

Value associated with this metric.

metricTimeStamp :: Int

The timestamp at which this metric was recorded.

metricStep :: Maybe Int

Step at which to log the metric.

data MetricLogger

Constructors

MetricLogger

Fields

metricLoggerRunId :: String

ID of the run under which to log the metric. Must be provided.

metricLoggerKey :: String

Name of the metric. This field is required.

metricLoggerValue :: Float

Double value of the metric being logged. This field is required.

metricLoggerTimeStamp :: Int

Unix timestamp in milliseconds at the time metric was logged. This field is required.

metricLoggerStep :: Maybe Int

Step at which to log the metric.

data MetricBatchLogger

Batch Logging

Constructors

MetricBatchLogger

Fields

mblRunId :: String

ID of the run to log under

mblMetrics :: [Metric]

Metrics to log. A single request can contain up to 1000 metrics, and up to 1000 metrics, params, and tags in total.

mblParams :: [Param]

Params to log. A single request can contain up to 100 params, and up to 1000 metrics, params, and tags in total.

mblTags :: Maybe [Tag]

Tags to log. A single request can contain up to 100 tags, and up to 1000 metrics, params, and tags in total.

type ModelVersionTag = Tag

Tag for a model version.

data ModelVersionStatus

Status of a Model

Constructors

PendingRegistration
Request to register a new model version is pending as server performs background tasks.

FailedRegistration
Request to register a new model version has failed.

Ready
Model version is ready for use.

data ModelVersion

Model Version

Constructors

ModelVersion

Fields

modelName :: String

Unique name of the model

modelVersion :: String

Model’s version number.

creationTimeStamp :: Int

Timestamp recorded when this model_version was created.

lastUpdatedTimeStamp :: Int

Timestamp recorded when metadata for this model_version was last updated.

modelUserId :: String

User that created this model_version.

currentStage :: String

Current stage for this model_version.

modelDescription :: String

Description of this model_version.

modelSource :: String

URI indicating the location of the source model artifacts, used when creating model_version

modelRunId :: String

MLflow run ID used when creating model_version, if source was generated by an experiment run stored in MLflow tracking server.

modelStatus :: ModelVersionStatus

Current status of model_version

modelStatusMessage :: String

Details on current status, if it is pending or failed.

modelTags :: Maybe [ModelVersionTag]

Tags: Additional metadata key-value pairs for this model_version.

modelRunLink :: String

Run Link: Direct link to the run that generated this version.

data Param

Param associated with a run.

Constructors

Param

Fields

paramKey :: String

Key identifying this param.

paramValue :: String

Value associated with this param.

type RegisteredModelTag = Tag

Tag for a registered model

data RegisteredModel

Constructors

RegisteredModel

Fields

regName :: String

Unique name for the model.

regCreationTimeStamp :: Int

Timestamp recorded when this registered_model was created.

regLastUpdatedTimeStamp :: Int

Timestamp recorded when metadata for this registered_model was last updated.

regUserId :: String

User that created this registered_model

regDescription :: String

Description of this registered_model.

latestVersions :: [ModelVersion]

Collection of latest model versions for each stage. Only contains models with current READY status.

regTags :: Maybe [RegisteredModelTag]

Tags: Additional metadata key-value pairs for this registered_model.

type RunTag = Tag

Tag for a run.

data RunStatus

Status of a run.

Constructors

Running
Run has been initiated.

Scheduled
Run is scheduled to run at a later time.

Finished
Run has completed.

Failed
Run execution failed.

Killed
Run killed by user.

data RunData

Run data (metrics, params, and tags).

Constructors

RunData

Fields

runMetrics :: Maybe [Metric]

Run metrics.

runParams :: Maybe [Param]

Run parameters.

runTags :: Maybe [RunTag]

Additional metadata key-value pairs.

data RunInfo

Metadata of a single run.

Constructors

RunInfo

Fields

runId :: String

Unique identifier for the run.

runExperimentId :: String

The experiment ID.

runStatus :: RunStatus

Current status of the run.

runStartTime :: Int

Unix timestamp of when the run started in milliseconds.

runEndTime :: Maybe Int

Unix timestamp of when the run ended in milliseconds.

runArtifactUri :: String

URI of the directory where artifacts should be uploaded. This can be a local path (starting with “/”), or a distributed file system (DFS) path, like s3:/bucketdirectory or dbfs:mydirectory. If not set, the local ./mlruns directory is chosen.

runLifeCycleStage :: LifeCycleStage

Current life cycle stage of the experiment : OneOf(“active”, “deleted”)

data Run

A single run.

Constructors

Run

Fields

runInfo :: RunInfo

Run metadata.

runData :: RunData

Run data.

data RunCreator

Request Structure for Creating Runs

Constructors

RunCreator

Fields

runExpId' :: String

ID of the associated experiment.

runStartTime' :: Int

Unix timestamp in milliseconds of when the run started.

runTags' :: Maybe [RunTag]

Additional metadata for run.

data RunUpdater

Request Structure for Updating Run metadata.

Constructors

RunUpdater

Fields

uRunId :: String

ID of the run to update. Must be provided.

uStatus :: RunStatus

Updated status of the run.

uEndTime :: Maybe Int

Unix timestamp in milliseconds of when the run ended.

data ViewType

View type for ListExperiments query.

Constructors

ActiveOnly
Default. Return only active experiments.

DeletedOnly
Return only deleted experiments.

All
Get all experiments.